requirements.txt
flask==3.0.2
langchain-groq==0.1.1
langchain-core==0.1.27
python-dotenv==1.0.1
gunicorn==21.2.0 